working on system security

+++ rdfframework/rdfclass.py
@@ -7,7 +7,8 @@ from werkzeug.datastructures import FileStorage
 from jinja2 import Template
 from rdfframework.utilities import clean_iri, fw_config, iri, is_not_null, \
     make_list, make_set, make_triple, remove_null, DeleteProperty, \
-    NotInFormClass, pp, uri, calculate_default_value, uri_prefix, nouri, get_attr
+    NotInFormClass, pp, uri, calculate_default_value, uri_prefix, nouri, \
+    pyuri, get_attr, slugify
 
 from .getframework import get_framework as rdfw
 from rdfframework.rdfdatatype import RdfDataType
@@ -123,7 +124,7 @@ class RdfClass(object):
             return self._run_save_query(save_query)
         #return None
 
-    def new_uri(self):
+    def new_uri(self, **kwargs):
         '''*** to be written ***
         generates a new URI
             -- for fedora generates the container and returns the URI
@@ -133,19 +134,23 @@ class RdfClass(object):
         else:
             debug = False
         if debug: print("START RdfClass.new_uri ---------------------------\n")
-        return_val = None
+        return_val = ''
         if self.kds_saveLocation == "triplestore":
             # get a uuid by saving an empty record to the repo and then
             # delete the item.
-            repository_result = requests.post(
-                            self.repository_url,
-                            data="",
-            				headers={"Content-type": "text/turtle"})
-            object_value = repository_result.text
-            uid = re.sub(r'^(.*[#/])','',object_value)
-            requests.delete(object_value)
-            if debug: print("new uid: ", uid)
-            return_val = self.uri_patterner(uid)
+            if "!--uuid" in self.kds_subjectPattern:
+                repository_result = requests.post(
+                                self.repository_url,
+                                data="",
+                				headers={"Content-type": "text/turtle"})
+                object_value = repository_result.text
+                uid = re.sub(r'^(.*[#/])','',object_value)
+                requests.delete(object_value)
+            else:
+                uid = ''
+            if self.kds_storageType != "blanknode":
+                if debug: print("new uid: ", uid)
+                return_val = self.uri_patterner(uid, **kwargs)
         return return_val
         if debug: print("END RdfClass.new_uri ---------------------------\n")
 
@@ -168,12 +173,23 @@ class RdfClass(object):
                     value = nouri(self.kds_classUri)
                 elif arg == "!--uuid":
                     value = uid
+                elif arg.startswith("!--slugify"):
+                    _prop_uri = arg[arg.find("(")+1:arg.find(")")]
+                    _prop_uri = pyuri(_prop_uri)
+                    data = kwargs.get('save_data',{}).get("data",[[]])
+                    _value_to_slug = ""
+                    for item in data:
+                        if item[0] == _prop_uri:
+                            _value_to_slug = \
+                                    item[1][1:item[1].find('"^^xsd')]
+                    if is_not_null(_value_to_slug):
+                        value = slugify(_value_to_slug)
                 else:
                     value = arg.replace("!--","")
             else:
                 value = arg
             new_args.append(value)
-        new_uri = "".join(new_args)
+        new_uri = uri("".join(new_args))
         if new_uri.startswith("http://"):
             temp_uri = new_uri.replace("http://","").replace("//","/")
             if temp_uri[:1] == "/":
@@ -407,7 +423,6 @@ class RdfClass(object):
             return self.kds_properties.get(prop_uri)
         except:
             return None
-
 
     def _validate_required_properties(self, rdf_obj, old_data):
         '''Validates whether all required properties have been supplied and
@@ -688,7 +703,7 @@ class RdfClass(object):
         if not DEBUG:
             debug = False
         else:
-            debug = False
+            debug = True
         if debug: print("START RdfClass._generate_save_query -------------\n")
         _save_data = save_data_obj.get("data")
         # find the subject_uri positional argument or look in the save_data_obj
@@ -702,7 +717,7 @@ class RdfClass(object):
             _save_type = "object"
         new_status = False
         if self.kds_saveLocation == "triplestore" and subject_uri == "<>":
-            subject_uri = iri(self.new_uri())
+            subject_uri = iri(self.new_uri(save_data=save_data_obj))
             new_status = True
         _bn_insert_clause = []
         _insert_clause = ""
